Created User, UserRepository, TestMain, created tests for Store, UserRepository

+++ migrations/20201225143356_create_users.up.sql
@@ -1,5 +1,5 @@
 create table users (
-    id int not null primary key,
+    id bigserial not null primary key,
     email varchar not null unique,
     enc_password varchar not null
 );
